Fix the issues with declaring the instantiation of an argument
My method for getting the created instantiations was relying on being sent a
var for generic var/const fields, but in the type constructor case we send in
a type.  Update my check for prior instantiations to accomodate this.

Also, collectInstantiatedAggregateTypes had issues when "instantiatedFrom" was
not the first generic type, so update that value when we obtain it via type
constructors (since I don't care about that field in my initializers
implementation otherwise)