Remove chpl-mem-warning-macros.h
chpl-mem-warning-macros.h was used to prevent accidental calls to the system
allocator in the runtime. It did so by redefining malloc and friends to bogus
values. This often caused build failures on cygwin because system headers were
updated and included the token "malloc", which prevented us from building:

 - https://github.com/chapel-lang/chapel/commit/95272793b4
 - https://github.com/chapel-lang/chapel/commit/59e3f5e5bb

We also had to turn off the warnings anytime we wanted to reference the token
malloc (such as in the jemalloc shim), which was kind of a pain.

We now have travis/jenkins jobs that run ./util/devel/lookForBalloc, so rely on
them instead of chpl-mem-warning-macros.h and chpl-mem-no-warning-macros.h.