Move chpl_getPrivatizedClass() into chpl-privatization.h
This is a second attempt at #6198, but only moves chpl_getPrivatizedClass()
instead of the entire privatization implementation. chpl_getPrivatizedClass()
is just a getter for chpl_privateObjects, so we also need to extern to
chpl_privateObjects.

chpl_getPrivatizedClass() can be called frequently, so we want to allow the
backend compiler to fully optimize/inline calls to it. This has a pretty big
performance impact for the stencil PRK, improving performance by about 15% for
16-node-xc. There's also some minor improvements for fft, and lulesh.

This is motivated by #6184, though it's not quite enough to close that issue
yet.