"Inline" record _multiData.
Replace the tuple of _multiData records in the _remoteAccessData record
and the DefulatRectangularArr class with tuples of its components.  This
seems to let us uncomment the "local field" pragmas without causing the
"error: Attempted to assign to local class field with remote class"
message we'd seen before with comm=gasnet.