Revert "Move runtime privatization support into chpl-privatization.h"
This reverts #6198. That PR attempted to move all the runtime privatization
support into a header so the backend could fully optimize/inline the
privatization code. In hindsight, this was a really broken commit because the
privatization array and syncvar are static globals in the file. We include
privatization in chpl-init.c to call chpl_privatization_init() and in stdchpl.h
so we can access it from the module code. Putting everything in the header
meant init.c and stdchpl.h had different arrays/syncvars, so chpl-init wasn't
initializing the one used by the modules.