Update the .bad file after my generic instantiation change
Prior to my #6209 to support type constructors and declarations of generic
instantiations, this test was failing with an internal compiler segfault.  As
part of my change, I added an error message with generic records that define
initializers, because the error mode they would encounter was not specifically
related to them rather than something due to them and sometimes due to the
type constructor issue.  Update the .bad file to reflect this new error mode.