Improve multi-ddata field and field accessor function names.
I wasn't fond of the md*Acc() accessor function names for the recently
"inlined" fields of the old _multiData record.  Change the names of the
fields to md*Vec, and change the names of the accessor functions to just
md*().

While here, also update several tests to reflect the new field and
accessor function names.