Fix a long-standing bug for scoped access to internal modules.
See also PR #4476 and PR #4517 which reverted it.  This PR makes a simple
change to findVisibleFunctions to fix a where scoped access to a module
provided automatically does not compile; for example

  writeln(List.makeList(1,2,3));

(with or without a 'use List').

findVisibleFunctions contains an optimization which is important for the
performance of the compiler. This optimization is to store all
included-by-default symbols into one scope instead of performing extra
bookkeeping for these scopes across all of the included-by-default code.
The code handling List.makeList (e.g.) simply needed to be updated to use
this one scope when appropriate.