Don't ignore running tasks in DefaultRectangular iterators.
For locModel!=flat we long ago switched to ignoring the running task
count for purposes of chunking iterations in DefaultRectangular.  But
since then we've changed how we compute the number of tasks to use in
nested iteration loops, so we actually shouldn't be ignoring other
running tasks any more.  Here, take into account other running tasks
when computing how many total tasks to apply to the nested looping
constructs we use in DefaultRectangular iteration when locModel!=flat.