Solve a couple of typos I spotted, and mention first class functions
Lydia's improvements to this draft chip
* Solve a couple of typos I spotted, and mention first class functions

Lists first class functions as an alternative for those functions that rely on
calling other functions not necessarily defined at their definition point.

* More tweaking to the public/private section

* Syntax unification and turn "CHIP 2" references into a link

* More formatting, ugh

* Sorry for the noise, trying to get the link right

* Give up on link

I think prior strategies that worked are currently broken, maybe due to the
interpreter changing for github.  Not going to mess with further.