Improve range iterator running task count accuracy, for locModel!=flat.
If we're told to take running tasks into account in an iterator, when
the locale model has sublocales, do so when computing the number of
tasks to use for the whole two-level loop nest, not separately for the
outer and inner loop.  Doing so for the inner loop especially isn't
right, since we can end up reducing our own task count because of the
existence of tasks created to run the outer loop.