Fix math in user's NPB MG
This program would fail when run with more than 4 locales. The cause
turned out to be some incorrect math in the 'stencilConvolve' helper.
This core operation is written with a temporary array in the reference
version, but we don't have task-private arrays yet so we wrote it
differently for Chapel. It seems like one or more aspects of the
computation were translated incorrectly in the original version of this
code.

After these changes, we can run with 8 and 16 locales without failures.
Performance for 4 nodes seems to be roughly the same as it was before
this change.

I also changed the coefficient type from an array to a tuple, and made
some minor whitespace changes.