Move the short policy section of "develop your contribution locally"
Makes it a section that follows "Create new branch" instead of burying it in
the HOWTO section, but leave most of the git commands there.  Move the branch
creation command out (so that section isn't so trivial).  Rename "develop your
contribution locally" to "development commands" to reflect that it is only
commands now.  Fix some typos ("every" -> "ever")