Migrate "read commit messages" section to HOWTO
Add a line mentioning it in the "Develop and test contributions locally"
section.  Add a link from it to "Other logging commands", after updating the
name of that section to match the capitalization scheme.  Remove blank section
I accidentally left over from moving "Create and switch to a feature branch"