Add test showing winner of dynamic dispatch vs. forwarding
While trying to use forwarding to support 'Replicated' and
'ReplicatedDist' without code duplication yesterday, I was thwarted by
the fact that a method that can either be dynamically dispatched or
forwarded will choose the dynamic dispatch case.  Michael argues that
this is intentional because the role of forwarding is to serve as a
last resort.  I can buy this argument, so am filing this test to lock
the current behavior in.