Add a bunch more tests
The directory classes/initializers/records/generics contains versions of the
tests from classes/initializers/generics that are modified to use the instance
declaration with only specified types on records.  This commit adds a version of
some of those tests which use the original initialization expression, to
validate those cases work as they are slightly less complicated (and don't run
into the question of "how many arguments should be sent with the default call
to the record initializer?").  6 of them work out of the box with my fix for
generic records with initializers, but more work is needed for the remaining 12