Remove 'where true' from ~ on bools
In PR #6300, I added a `where` clause to proc ~ for boolean types in the spirit
of making it overloadable by user code.  Unfortunately I had that backwards - a
function is *more specific* if a where clause matches, not the other way
around. Anyway, the test of user overload of proc ~ for bools still works
because the proc ~ it finds is scoped closer than the standard modules.