Update .bad file for this future, after checking with Brad
My change to support initializers on generic records allowed this future to
progress further in compilation than it was previously able to.  However, now
we encounter an issue when accessing the contents of the array field - it seems
as though perhaps initializers are not storing the result of an array creation
expression (`[i in 1..10] i;`) properly, so we get "error: illegal access of
iterator or promoted expression" when we try `field[idx]`

This bears further investigation, but I am going to focus on other things for
now.