Add a future to track the type of iterable expressions in multiple contexts
I discovered that the type of an iterable expression is different when it is
saved into a local variable versus when it is passed as an argument to a
function.  This future tracks that difference, and poses the question of
whether the types should match and what it should be if they do.