Suppress known LLVM issues (#6340)
Suppress known LLVM issues (#6340)

This is a trivial change adding four .suppressif files for CHPL_LLVM==llvm.  Positive and negative tested but not reviewed.  The idea was preapproved by @mppf .

This resulted from discussions with @mppf regarding the test failures documented in issues #6336 and #6338 . Since it will be a while before these can be fixed, adding .suppressif files will allow a clean paratest with LLVM for now, so that people making changes that affect LLVM can know that these issues are not their fault.