Modify copy initializer for generic records test
Now that generic records with initializers are partially supported, we started
running into an ambiguity in this test because the two initializers basically
had the same signature.  Fix that by declaring the type of the argument to the
second one to be explicitly R.  Unfortunately, there are still other issues with
generic records and copy initializers - even though the signature is correct,
the implementation is still not recognizing that the second initializer is a
copy initializer.  This will need to be fixed independent of this work.