Switch ResolutionCandidate's actual and formal tracking to use maps
Prior to this change, the fields formalIdxToActual and actualIdxToFormal were
sparsely populated Vecs.  This change modifies them to be std::maps.  There was
some reliance on the sparsely populated Vecs having a larger size than they
would necessarily be populated with, so I modified some iterations through these
structures to account for that change.