Create wrappers for generic record initializers
We had skipped performing this action during normal candidate elimination due to
the type of the "this" actual still being generic until the body had been
resolved.  With classes, this was fine because the initializer call would be
wrapped by the "_new" function and that function would get the appropriate
wrappers created for it (and doing so in both places would be unnecessary and
problematic).  Records don't define a "_new" function, though, so needed to
have their wrappers created after the concrete type had been determined, in
order to handle arguments with default values, etc.

This commit adds that support by taking the relevant portions of
computeActualFormalAlignment and wrapAndCleanUpActuals and adjusting them to
suit this specific use case.