Add a multi-locale performance test to track remote task startup
Add a performance test to track the cost of doing:

   coforall loc in Locales do on loc {}

This idiom is a fundamental building block of our distributions and parallel
applications, so we should track its performance. This is similar to the
task-spawning benchmarks added in #4102, except that we're spawning remote
tasks instead of local ones.

We've recently discovered that remote task spawning for coforall+on's is more
expensive when using network atomics (#6356), so this test will serve as a way
to track progress on optimizations that improve that and serve as a baseline
comparisons to tree-based spawning schemes we want to explore in the future.