Allow 'local field' pragma on arrays
The optimized Stencil PRK currently relies on a local block to squash
wide-pointer overhead. At a high-level, this should be done instead by
applying the 'local field' pragma to certain fields in the various
StencilDist classes. For the pragma to have effect, a change is made
to insertWideReferences to allow 'local field' to work on
PRIM_GET_MEMBER. A change was also made to squash the warning for arrays.