First attempt at fixing `var x: MyRec(generic args);` for initializers
This is a proof of concept commit.  I was able to get 14 of the generic record
with initializers futures related to instance declarations with type
information to work due to this change (and it is likely that some of the
remaining 5 were because I did something dumb when updating the test).

Basically, it turns off the error message for calling the type constructor of
generic records with initializers and copies some of the logic used by the
_defaultOf call to the default constructor when creating the _defaultOf function
for generic records with initializers.  The main difference is that the
arguments inserted are not "named" arguments and that the init call is not
treated as though it will return a result (because it won't).