Test improvements due to the previous commit
Remove the .future files on the 15 new passing futures and the .bad file.  Also,
fix the expected output for one test and remove an accidentally leftover line
in another so that they are both capable of passing as expected.

With this change, only one future test remains that should be addressed as this
set of operations go in
(classes/initializers/records/generics/explicitAssignment.chpl, which isn't
following the right rules for some bit of the operations).