Share handling of var fields with constructors
The code I added was very similar to the handling of the var field arguments to
the default constructors, with the only major differences being that I didn't
try to handle "outer" specially and that I don't insert the actual in a named
form.  This commit changes my implementation so that I am sharing the code
with the constructors implementation.