Add a pragma "function never returns" for functions like 'halt' and 'exit'
Add a pragma for functions like 'halt' and 'exit' that don't return but cause
the program to terminate instead. Check for this pragma in returns-on-all-paths
analysis instead of checking for functions named 'halt'.

Added the new pragma to the definitions of 'halt' and 'exit'.

Remove 3 .future and .bad files for newly passing tests.