Testing updates for this change
With the previous commit, the record version of this test now behaves the same
as a generic record with constructors.  The class version was already consistent
and upon further reflection, I understood the reason why classes and records
would and should behave differently in this regard, so I decided to update the
expected output for the class version and remove the .future.