Revert this check.
When I made my first attempt to support `var x = new MyRec(int);`, I thought
that I could update the handling of the typed version in the same way.  However,
the typed version won't fall into that category, because the type expression
is too complicated with all the generic field arguments.  Undo the change I
made to support it, since it won't be useful.