Enable testing LCALS in correctness testing
LCALS has only been getting tested for performance testing because with bounds
checks it takes way to long to run. Add a compopts file with
'--no-bounds-checks' to enable it to run in regular correctness testing.

Modified the test config files to run the large loop size for the RAW and
RAW_OMP variants.

Modified a config param to allow selecting between three output formats:
* 0: Don't print timing/checksum dumps
* 1: Like the reference code
* 2: Formatted nicely for perf testing

I tried running with bounds checking, and hit an out of bounds error at around
the 40 minute mark. Looking into this, I found that the reference version also
accesses the array out of bounds with index -1.  I added a comment mentioning
this where it happens.