Implement an optimized variant of updateFluff()
The original updateFluff implementation might end up using many GETs or
PUTs for noncontiguous chunks of memory. This commit implements a
variant of updateFluff that packs the elements into a buffer to be used
in a single GET or PUT, and is unpacked on the destination locale. This
optimization will fire if the element types can be bulk-transferred.

To implement this change, the StencilDom.setup() method was modified
such that each locale now knows how to send its own indices or how to
fetch indices. This information is stored arrays of domains which are
used in the optimized updateFluff method.

This optimized variant can be disabled with the new config param:
  stencilDistAllowPackedUpdateFluff

This optimization was motivated by the Stencil PRK when we observed that
our updateFluff stage was performing significantly worse than the
reference's corresponding stage. This commit improves our performance
depending on the comm layer used. In gasnet-mpi, we see a 10x
improvement. In ugni-qthreads, a 3x improvement.

Minor style cleanups were made while working on this change.