Commit updated version of default initializers change
When building the default constructor for an AggregateType (which currently
is called during scopeResolve.cpp), determine if the type is such that we
should create a default initializer.  This is currently true if:
- The type is not defined in an internal or standard/package module
- The type does not have a user-defined initializer
- The type does not have a user-defined constructor
- The type is not a nested type
- The type does not have an "initialize" method
- All parent types also fulfill these conditions.

Change a couple of checks against DEFINES_INITIALIZER to also include the
DEFINES_NONE_USE_DEFAULT case, at least when it is a default initializer.

Add a flag `--force-initializers` which will enable the use of default
initializers for certain types in user code.

Add a function to build the default initializer to buildDefaultFunctions,
which will currently only fire for user classes that aren't nested and don't
have an initialize method.

Add an error when we wanted a default initializer defined but for some reason
weren't able to.