Repair the array-loc test.
Modify the array-loc.suppressif file so that it only suppresses a
failure when libnuma was linked in, which we take as indicating that we
should have tried to localize the array memory.  For the time being, in
this case the test fails because we can't actually localize without the
multi-ddata feature, which has been mothballed.  The test doesn't fail
if libnuma isn't linked in, because in that case it reports "unknown"
locality and expects the same.

Also, adjust the code that computes the memory address whose locality we
key off of.  This was broken by #6060 though I didn't realize it at the
time.  Note, though, that this fix is only needed for the multi-ddata
case, so I'm actually cleaning up unused code here