Add a very simple test of --print-module-resolution
Grepping tests, I couldn't find any existing tests that
throw --print-module-resolution. So here I add one just
to get it tested at all.

The prediff for this test takes drastic measures to avoid it
becoming a maintenance burden. It only prints out the line
for the tested module (because included-by-default modules can change)
and changes all the numbers to nnnn (because number of ASTs in
a simple module can change).