Simplify dsiAccess() interface; reverse haveLock sense in findFilledSlot()
On Michael's recommendation, removed the haveLock argument from the
dsiAccess() interface (because it was effectively unused) and
reversed 'haveLock' in findFilledSlot() to 'needLock' because it's
a better description of what we want (We may not have the lock, but
also may not need it.  In the old world, this amounted to 'haveLock=true'
even if we didn't actually hold a lock).