Prepare PRIM_ARRAY_FREE for network-aware allocation.
To support NUMA locality in NIC-registered memory, at least the ugni
comm layer will soon provide dynamic array memory registration.  This
will require allocating and freeing registerable memory for "large" (to
be defined later) arrays in the comm layer.  In turn, specifically to
support freeing such memory efficiently, we need to be able to tell
whether a given array's memory came from the comm layer or the memory
layer.  Here we require _ddata_free() calls to repeat the element count
and element type size given to the corresponding _ddata_allocate(), so
that we can distinguish between large and small arrays in the free()
code just as we do in the allocate() code.  So far this size information
is piped all the way down to the runtime, but as yet we aren't making
use of it.  That's the next step.

While here, I did some related cleaning up.  The eltType argument to
PRIM_ARRAY_ALLOC was unused, so I removed it.  And PRIM_ARRAY_FREE_ELTS
was also unused, so I removed it as well.

NOTE: currently this does not quite work.  For the comm=gasnet (wide
pointer) case I am getting compilation failures on the emitted C code
for the PRIM_ARRAY_FREE invocations.  I've enlisted help to solve this.