Add some tests of inheriting from generic classes
I am covering the interactions with default initializers, but realized that I
didn't have tests of the parent/child relationship when both have initializers
and the parent is generic.  Unfortunately, it doesn't currently work (and
default initializers won't solve it, but will probably run into the same issues
so I might solve it for both at the same time).