Add options to print unused functions
Add options --print-unused-functions and --print-unused-internal-functions.
The first prints the names and source locations of any unused functions in
the user's program.  The second is a developer option, and prints names
and source locations of unused functions across the entire program,
including in the modules.

The printing happens near the end of the function resolution pass, right
before unused functions are pruned away.

Added a small test that contains an unused function and uses the
--print-unused-functions flag.