Fix buildDefaultInitializer somewhat
Remove some commented out code that will never run, because it isn't useful.
Change the "else if parent defines an initializer" branch to be an assert in
just a plain else branch, as the "defines a constructor" case won't happen in
that nested branch.  Move the error that was for the "defines a constructor"
case to where it will actually be encountered, instead of just skipping in favor
of the general "we couldn't create that initializer" message.

Update tests that relied on that output.