Improve documentation for associative domains/arrays w.r.t. parallel safety
This PR adds a general description of parallel safety w.r.t. domains
and arrays to the user-facing online array documentation and also
documents the 'parSafe' option for associative domains.  This
content should ultimately go into the spec, but my sense is that we
are not confident at present that we have reached the final word
on this topic.

It also adds a short mention to the spec's description of extending
1:1 associative arrays via indexing to indicate that it is not
parallel-safe.