Emit warning suppression flag in the correct order (#6480)
Emit warning suppression flag in the correct order (#6480)

For a long time, we have been emitting gcc's -Wno-strict-overflow flag when compiling the generated code, but we never noticed until gcc 7 that it was placed on the command line in the wrong order. We had been putting it before -Wall which caused it to be overridden. This change places it after -Wall.

This fixes the test modules/packages/LinearAlgebraJama/TestMatrix when run with --fast.