Free by default
On master since the dawn of time, and in my previous commit, SSCA2 by
default would not clean up after itself.  Here, I flipped the sense of
that flag so that it would clean up after itself because I couldn't
stand the thought of editing more .compopts files when I realized that
the study versions share that same code.  I also noted that the study
versions had their own versions of the cleanup routine, so I added the
same bug fix to them that I did on the release version in my previous
commit.