Use == instead of strcmp on astr'ed strings
One of the motivations for using astr() is to allow comparing strings
using pointer equality (on canonicalized pointers) instead of strcmp().
This change leverages that in a few cases that are executed frequently
during compilation.

* These globals contain canonicalized strings as follows:

astrSdot    "."
astrSequals "="
astr_cast   "_cast"
astrDeinit  "deinit"
astrTag     "tag"
astrThis    "this"

(I introduced "tag" because I'd like to use it in an upcoming bug fix.)

(The S in astrSdot et al. stands for a "symbol", to distinguish from a word.)

* Most comparisons against those strings are replaced with ==.

* Also added isNamedAstr() as a companion to isNamed()
for use when the argument is canonicalized.
Engaged it is most cases where it fits.

I did not switch to using astrXXX globals for uses like

  new CallExpr(".", ...)

because (a) these occur a lot in the source files, probably not as much
during compilation, (b) it does not save us from a hashtable lookup
for each occurrence.

While there, fixed NamedExpr::name to canonicalize it as well.