Couple of clean ups
Only build the default initializer for the parent if wantsDefaultInitializer is
true (which prevents issues where a user class inherits from a standard module
class and so forces that class to use a default initializer when we aren't
necessarily ready for that to happen).

Also, only build the class allocator ("_new") function for classes.  This was
true today, but only because we don't build default initializers for records
yet.