Make --force-initializers a developer-only flag for now
I think that if a user runs into an issue with default initializers, we might
want to point them to this flag to turn default constructors back on and see
if the problem persists.  However, it is not on by default right now and I'm not
sure when it will be, so make it developer-only for now.