Fix deinitialization order in AutoDestroyScope::variablesDestroy
Variables need to be deinitialized in the reverse declaration order
(because they could refer to each other - or memory managed by each
other). Here I add a new test that checks deinitialization order in a way
that is meant to trigger different behavior in
AutoDestroyScope::variablesDestroy. Indeed, this new test reveals a
problem with the deinitialization order of variables declared in simple
nested blocks.

Then I fix AutoDestroyScope::variablesDestroy to deinitialize such
examples in the correct order. I did that by adding a PRIM_NOOP at the
end of blocks without a terminating GOTO; then in any case the
autoDestroy calls are added in reverse declaration order. This PRIM_NOOP
is added simply as a placeholder that the following code can insert
before and it is removed after it is no longer useful for this purpose.