Automatically rebuild third-party libraries
We're pretty good at rebuilding the compiler/runtime when code changes, but we
never rebuild the third-party libraries. When we upgrade third-party versions
or start throwing different configuration options we don't do anything to
rebuild the third-party library. This either means that people aren't
testing/using the new version, or for breaking changes they have to do a
clobber to fix the build. This often confuses users (internal and external.)

This adds code to automatically rebuild third-party libraries if any of their
C/C++ source files or our makefiles change.

Ideally we would also rebuild if any of the third-party files change, but
unfortunately for several packages that use autoconf such as gasnet, hwloc, and
qthreads we touch some configure/autoconf files in a certain order to prevent
autoconf from running, and this changes the timestamps, which causes make to
think the files were modified and we need to rebuild.

Note: test-venv and chpl-venv already auto-rebuild since they depend on
CHPL_VENV_*_REQUIREMENTS_FILE

Note: not auto-rebuildling llvm or fltk since they are force targets in
third-party/Makefile. Probably doesn't matter too much for fltk, but will
probably be important for llvm since we want to start pushing for it to be our
default.

TODO: try to rebuild if envs like CHPL_QTHREAD_ENABLE_OVERSUBSCRIPTION
CHPL_QTHREAD_MORE_CFG_OPTIONS, or CHPL_QTHREAD_NO_CHPL_ALLOC are changed?