Update test/arrays/vass/const-checking.field.chpl for initializers
Change the constructors in this test into initializers.  Also changed the
spacing slightly so that it is more legible.  The necessary addition of the
super.init(), and the addition of the line spacing meant that the expected
output needed slight modification.

I chose to put the end of Phase 1 prior to the array updates.  This means that
they violate some const-ness checking rules, but we don't detect that just yet.
Since the test is covering cases where we are supposed to detect const
modification, I think that was the right call, but it implies that we probably
need to improve our const-checking in Phase 2.