Update test/classes/bradc/test-class-init-blc.chpl for initializers
Converts the constructor in this test into an initializer.  Does this by adding
a super.init() call prior to the setting of the array field's contents.

Brad is okay with this update, but suggested that we should alter omitted field
initialization so that it would be valid to place the for loop to update the
array field's contents in Phase 1 (instead of Phase 2 where I put it).  That
would not prevent the double initialization of the contents (as it would still
give each index the default value for the type, it would just be altered in
Phase 1 instead of altered in Phase 2) - to do that, we would want to noinit the
array field and then update the contents.

The behavior of this test has changed, but we believe it is an improvement.
Before, the array field would be of size 0 and have no contents; now, it is of
length 3 with the appropriate value in each field.